src/backend/pyproject.toml
[project]
name = "mingai-backend"
version = "1.0.0"
description = "mingai Enterprise RAG Platform - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0,<0.28",
    "structlog>=24.1.0",
    "bcrypt>=4.1.0",
    "openai>=1.10.0",
    "cachetools>=5.3.0",
    "tenacity>=8.2.0",
    "psycopg[binary]>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "Pillow>=10.0.0",
    "slowapi>=0.1.9",
    "jsonschema>=4.20.0",
    "sendgrid>=6.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
timeout = 10
markers = [
    "unit: Tier 1 unit tests (mocking allowed)",
    "integration: Tier 2 integration tests (real infra, NO mocking)",
    "e2e: Tier 3 end-to-end tests (real browser + backend, NO mocking)",
]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
